+++ src/components-v2/common/DepositFeesInformation.tsx
@@ -1,0 +1,70 @@
+import React from 'react';
+import { Button, Divider, Grid, Typography } from '@material-ui/core';
+import ArrowRightAltIcon from '@material-ui/icons/ArrowRightAlt';
+import { makeStyles } from '@material-ui/core/styles';
+
+const useStyles = makeStyles((theme) => ({
+	root: {
+		padding: theme.spacing(3),
+	},
+	title: {
+		fontSize: 18,
+	},
+	backButton: {
+		display: 'flex',
+		alignItems: 'center',
+		textTransform: 'none',
+	},
+	content: {
+		marginTop: theme.spacing(2),
+	},
+	divider: {
+		width: '100%',
+		margin: theme.spacing(2, 0),
+	},
+	arrowIcon: {
+		marginRight: 4,
+		fontSize: 16,
+		transform: 'rotate(-180deg)',
+	},
+}));
+
+interface Props {
+	onBackClick: () => void;
+}
+
+export const DepositFeesInformation = ({ onBackClick }: Props): JSX.Element => {
+	const classes = useStyles();
+
+	return (
+		<Grid container className={classes.root}>
+			<Grid container justify="space-between" alignItems="center">
+				<Typography variant="body1" display="inline" className={classes.title}>
+					Fee Information
+				</Typography>
+				<Button color="primary" onClick={onBackClick} className={classes.backButton}>
+					<ArrowRightAltIcon className={classes.arrowIcon} />
+					Back
+				</Button>
+			</Grid>
+			<div className={classes.content}>
+				<Typography variant="body1" color="textSecondary">
+					Fees are charged on certain Sett Vaults, and are sent to the BadgerDAO treasury, or shared with the
+					strategist who wrote the vault strategy.
+				</Typography>
+				<Divider className={classes.divider} />
+				<div>
+					<Typography variant="body2" color="textSecondary">
+						- Performance Fees are assessed on the profit that the strategy creates
+					</Typography>
+					<Typography variant="body2" color="textSecondary">
+						- Management Fees are charged on the entire principal over the course of a year
+					</Typography>
+					<Typography variant="body2" color="textSecondary">
+						- Both Performance and Management Fees are factored into all ROI calculations
+					</Typography>
+				</div>
+			</div>
+		</Grid>
+	);
+};