+++ src/components-v2/leaderboard/LeaderboardAccountInformation.tsx
@@ -7,7 +7,6 @@ import { Skeleton } from '@material-ui/lab';
 import BoostSuggestion from './BoostSuggestion';
 import NativeRankSuggestion from './NativeRankSuggestion';
 import ViewBoostButton from './ViewBoostButton';
-import { FLAGS } from '../../config/constants';
 
 const useStyles = makeStyles((theme) => ({
 	justifyCenterOnMobile: {
@@ -18,7 +17,7 @@ const useStyles = makeStyles((theme) => ({
 	boostContainer: {
 		[theme.breakpoints.down('xs')]: {
 			textAlign: 'center',
-			marginBottom: FLAGS.BOOST_OPTIMIZER ? 16 : 0,
+			marginBottom: 16,
 		},
 	},
 	headerValueText: {
@@ -67,13 +66,11 @@ const LeaderboardAccountInformation = observer(
 						</Typography>
 					</Grid>
 				</Grid>
-				{FLAGS.BOOST_OPTIMIZER && (
-					<div className={classes.infoContainer}>
-						<ViewBoostButton />
-						<BoostSuggestion />
-						<NativeRankSuggestion />
-					</div>
-				)}
+				<div className={classes.infoContainer}>
+					<ViewBoostButton />
+					<BoostSuggestion />
+					<NativeRankSuggestion />
+				</div>
 			</Grid>
 		);
 	},