+++ src/components/IbBTC/Redeem.tsx
@@ -234,6 +234,7 @@ export const Redeem = observer((): any => {
 							<ErrorText variant="subtitle1">
 								<span>A maximum of </span>
 								<Tooltip
+									enterTouchDelay={0}
 									className={classes.maxAmount}
 									title="Apply limit"
 									arrow
@@ -266,6 +267,7 @@ export const Redeem = observer((): any => {
 						<Grid item xs={6}>
 							<EndAlignText variant="body1">
 								<Tooltip
+									enterTouchDelay={0}
 									enterDelay={0}
 									leaveDelay={300}
 									arrow