+++ src/components/Sidebar/index.tsx
@@ -16,13 +16,14 @@ import {
 } from '@material-ui/core';
 import { makeStyles } from '@material-ui/core/styles';
 import { ExpandMore } from '@material-ui/icons';
-import { SITE_VERSION, NETWORK_LIST, FLAGS } from 'config/constants';
+import { SITE_VERSION, NETWORK_LIST } from 'config/constants';
 import NetworkWidget from 'components-v2/common/NetworkWidget';
 import { Route } from 'mobx-router';
 import { RootStore } from 'mobx/RootStore';
 import clsx, { ClassValue } from 'clsx';
 import SecurityIcon from '@material-ui/icons/Security';
 import { sidebarPricingLinks } from 'config/ui/links';
+import { FLAGS } from 'config/environment';
 
 const DRAWER_WIDTH = 240;
 
@@ -201,7 +202,7 @@ export const Sidebar = observer(() => {
 		<div className={classes.contentRoot}>
 			<List>
 				<ListItem button className={classes.listItem}>
-					<img alt="Badger Logo" src={'assets/badger-logo.png'} className={classes.logo} />
+					<img alt="Badger Logo" src={'/assets/badger-logo.png'} className={classes.logo} />
 					<ListItemSecondaryAction>
 						<NetworkWidget />
 					</ListItemSecondaryAction>
@@ -253,7 +254,7 @@ export const Sidebar = observer(() => {
 					}}
 				>
 					<ListItemIcon>
-						<img alt="Badger Setts Logo" className={classes.icon} src={'assets/sidebar/sett.png'} />
+						<img alt="Badger Setts Logo" className={classes.icon} src={'/assets/sidebar/sett.png'} />
 					</ListItemIcon>
 					<ListItemText primary="Sett Vaults" />
 				</ListItem>
@@ -277,7 +278,7 @@ export const Sidebar = observer(() => {
 							<ListItemIcon>
 								<img
 									alt="Badger Digg Icon"
-									src={'assets/sidebar/digg-white.png'}
+									src={'/assets/sidebar/digg-white.png'}
 									className={classes.icon}
 								/>
 							</ListItemIcon>
@@ -291,7 +292,7 @@ export const Sidebar = observer(() => {
 							<ListItemIcon>
 								<img
 									alt="Interest Bearing Badger Bitcoin Icon"
-									src={'assets/sidebar/ibbtc-white.svg'}
+									src={'/assets/sidebar/ibbtc-white.svg'}
 									className={classes.icon}
 								/>
 							</ListItemIcon>
@@ -320,7 +321,7 @@ export const Sidebar = observer(() => {
 								}}
 							>
 								<ListItemIcon>
-									<img alt="Boosts" src={'assets/sidebar/boosts.png'} className={classes.icon} />
+									<img alt="Boosts" src={'/assets/sidebar/boosts.png'} className={classes.icon} />
 								</ListItemIcon>
 								<ListItemText primary="Boost" />
 								<IconButton
@@ -333,16 +334,14 @@ export const Sidebar = observer(() => {
 							</ListItem>
 						)}
 						<Collapse in={expanded === 'boosts'} timeout="auto" unmountOnExit>
-							{FLAGS.BOOST_OPTIMIZER && (
-								<ListItem
-									button
-									classes={{ gutters: classes.subItemGutters }}
-									className={getItemClass('/boost-optimizer', classes.primarySubListItem)}
-									onClick={() => navigate(views.boostOptimizer)}
-								>
-									Boost Optimizer
-								</ListItem>
-							)}
+							<ListItem
+								button
+								classes={{ gutters: classes.subItemGutters }}
+								className={getItemClass('/boost-optimizer', classes.primarySubListItem)}
+								onClick={() => navigate(views.boostOptimizer)}
+							>
+								Boost Optimizer
+							</ListItem>
 							<ListItem
 								button
 								classes={{ gutters: classes.subItemGutters }}
@@ -366,7 +365,7 @@ export const Sidebar = observer(() => {
 							<ListItemIcon>
 								<img
 									alt="Badger Arcade"
-									src={'assets/sidebar/gas_station.png'}
+									src={'/assets/sidebar/gas_station.png'}
 									className={classes.icon}
 								/>
 							</ListItemIcon>
@@ -421,7 +420,7 @@ export const Sidebar = observer(() => {
 					}
 				>
 					<ListItemIcon>
-						<img alt="Nexus Logo" src={'assets/sidebar/nexus_logo_bw.png'} className={classes.icon} />
+						<img alt="Nexus Logo" src={'/assets/sidebar/nexus_logo_bw.png'} className={classes.icon} />
 					</ListItemIcon>
 					<ListItemText>
 						Get Coverage